refactor(init): rename org prefetch helper
